refactor(web): split App.tsx into smaller components
App.tsx (649行) を 400 行ガイドライン以内に分割する pure refactor。

分割マップ:
- lib/routing.ts: readFromUrl/buildSearch/readFeedDetailFromPath/readAuthorFromPath/readViewFromUrl/formatRelative を抽出
- hooks/useFilterHandlers.ts: pushFilter と全フィルタハンドラ群を抽出
- components/EmptyState.tsx: EmptyState コンポーネントを抽出
- components/StaleFeedsWarning.tsx: stale feeds の details セクションを抽出
- components/ArticlesView.tsx: articles ビュー全体 (toolbar + search + filter + list + load-more) を抽出
- App.tsx: routing/kb/navigation ロジックのみ残す

行数 before/after:
  App.tsx: 649 -> 283
  ArticlesView.tsx: (新規) 256
  useFilterHandlers.ts: (新規) 207
  lib/routing.ts: (新規) 78
  StaleFeedsWarning.tsx: (新規) 30
  EmptyState.tsx: (新規) 17

Closes #257
